frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/validation/EmailCodeProcessor.java
package uk.gov.di.authentication.frontendapi.validation;

import uk.gov.di.authentication.shared.entity.ErrorResponse;
import uk.gov.di.authentication.shared.services.AuditService;
import uk.gov.di.authentication.shared.services.AuthenticationService;
import uk.gov.di.authentication.shared.services.CodeStorageService;
import uk.gov.di.authentication.shared.services.ConfigurationService;
import uk.gov.di.authentication.shared.services.DynamoAccountModifiersService;
import uk.gov.di.authentication.shared.state.UserContext;

import java.util.Optional;

public class EmailCodeProcessor extends MfaCodeProcessor {

    EmailCodeProcessor(
            CodeStorageService codeStorageService,
            UserContext userContext,
            ConfigurationService configurationService,
            AuthenticationService dynamoService,
            AuditService auditService,
            DynamoAccountModifiersService dynamoAccountModifiersService) {
        super(
                userContext,
                codeStorageService,
                configurationService.getCodeMaxRetries(),
                dynamoService,
                auditService,
                dynamoAccountModifiersService);
    }

    @Override
    public Optional<ErrorResponse> validateCode() {
        return Optional.empty();
    }

    @Override
    public void processSuccessfulCodeRequest(String ipAddress, String persistentSessionId) {}
}
