frontend-api/src/test/java/uk/gov/di/authentication/frontendapi/validation/EmailCodeProcessorTest.java
package uk.gov.di.authentication.frontendapi.validation;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import uk.gov.di.authentication.frontendapi.domain.FrontendAuditableEvent;
import uk.gov.di.authentication.shared.entity.Session;
import uk.gov.di.authentication.shared.services.AuditService;
import uk.gov.di.authentication.shared.services.AuthenticationService;
import uk.gov.di.authentication.shared.services.CodeStorageService;
import uk.gov.di.authentication.shared.services.ConfigurationService;
import uk.gov.di.authentication.shared.services.DynamoAccountModifiersService;
import uk.gov.di.authentication.shared.state.UserContext;

import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import static uk.gov.di.authentication.shared.entity.NotificationType.VERIFY_EMAIL;
import static uk.gov.di.authentication.shared.services.AuditService.MetadataPair.pair;

public class EmailCodeProcessorTest {
    private EmailCodeProcessor processor;

    private static final String TEST_CLIENT_EMAIL =
            "testclient.user1@digital.cabinet-office.gov.uk";
    private final AuditService auditService = mock(AuditService.class);
    private final CodeStorageService codeStorageService = mock(CodeStorageService.class);
    private final UserContext userContext = mock(UserContext.class);
    private final AuthenticationService authenticationService = mock(AuthenticationService.class);
    private final ConfigurationService configurationService = mock(ConfigurationService.class);
    private final DynamoAccountModifiersService accountModifiersService =
            mock(DynamoAccountModifiersService.class);
    private final Session session = mock(Session.class);

    @BeforeEach
    void setup() {
        when(session.getEmailAddress()).thenReturn(TEST_CLIENT_EMAIL);
        when(userContext.getSession()).thenReturn(session);
        processor =
                new EmailCodeProcessor(
                        codeStorageService,
                        userContext,
                        configurationService,
                        authenticationService,
                        auditService,
                        accountModifiersService);
    }

    @Test
    void shouldReturnNullWhenCorrectEmailCodeProcessed() {
        processor.validateCode();
        verify(auditService)
                .submitAuditEvent(
                        FrontendAuditableEvent.CODE_VERIFIED,
                        AuditService.UNKNOWN,
                        AuditService.UNKNOWN,
                        AuditService.UNKNOWN,
                        AuditService.UNKNOWN,
                        AuditService.UNKNOWN,
                        "123.123.123.123",
                        AuditService.UNKNOWN,
                        AuditService.UNKNOWN,
                        pair("notification-type", VERIFY_EMAIL.name()));
    }
}
